+++ apps/ensindexer/src/plugins/lineanames/lineanames.plugin.ts
@@ -6,6 +6,7 @@ import {
   makePluginNamespace,
   networkConfigForContract,
   networksConfigForChain,
+  validateContractConfigs,
 } from "@/lib/plugin-helpers";
 import { DatasourceName } from "@ensnode/ens-deployments";
 import { PluginName } from "@ensnode/utils";
@@ -21,18 +22,15 @@ export const requiredDatasources = [DatasourceName.Lineanames];
 const { chain, contracts } = MERGED_ENS_DEPLOYMENT[DatasourceName.Lineanames];
 const namespace = makePluginNamespace(pluginName);
 
+// validate runtime contract config
+validateContractConfigs(pluginName, contracts);
+
 export const config = createConfig({
   networks: networksConfigForChain(chain),
   contracts: {
     [namespace("Registry")]: {
       network: networkConfigForContract(chain, contracts.Registry),
       abi: contracts.Registry.abi,
-    },
-    [namespace("Resolver")]: {
-      network: networkConfigForContract(chain, contracts.Resolver),
-      abi: contracts.Resolver.abi,
-      // index Resolver by event signatures, not address
-      filter: contracts.Resolver.filter,
     },
     [namespace("BaseRegistrar")]: {
       network: networkConfigForContract(chain, contracts.BaseRegistrar),
@@ -46,18 +44,20 @@ export const config = createConfig({
       network: networkConfigForContract(chain, contracts.NameWrapper),
       abi: contracts.NameWrapper.abi,
     },
+    Resolver: {
+      network: networkConfigForContract(chain, contracts.Resolver),
+      abi: contracts.Resolver.abi,
+    },
   },
 });
 
 export const activate = activateHandlers({
   pluginName,
-  // the shared Registrar handler in this plugin indexes direct subnames of '.linea.eth'
-  registrarManagedName: "linea.eth",
   namespace,
   handlers: [
     import("./handlers/Registry"),
     import("./handlers/Registrar"),
-    import("./handlers/Resolver"),
     import("./handlers/NameWrapper"),
+    import("../shared/Resolver"),
   ],
 });