+++ packages/ponder-subgraph-api/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./middleware";
-export * from "./types";
-export { buildGraphQLSchema } from "./graphql";