+++ packages/ens-deployments/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ensnode/ens-deployments
 
+## 0.26.0
+
+## 0.25.0
+
+### Minor Changes
+
+- [#657](https://github.com/namehash/ensnode/pull/657) [`edd2b1e`](https://github.com/namehash/ensnode/commit/edd2b1ebb4e052a5036edee090dd05c80cb732ca) Thanks [@shrugs](https://github.com/shrugs)! - added threedns datasources
+
+## 0.24.0
+
+### Minor Changes
+
+- [#662](https://github.com/namehash/ensnode/pull/662) [`ea87355`](https://github.com/namehash/ensnode/commit/ea87355e2893448bf53d586bef436ce20095b66d) Thanks [@shrugs](https://github.com/shrugs)! - update ens-test-env to read from env variables for deployment addresses
+
 ## 0.23.0
 
 ## 0.22.1